Refactor code and add Dependency class
