+++ Unosquare.Labs.SshDeploy/Utils/Dependency.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Unosquare.Labs.SshDeploy.Utils
+{
+    class Dependency
+    {
+        public string Name { get; set; }
+        public string Version { get; set; }
+        public string Path { get; set; }
+    }
+}
